+++ src/src_user/IfWrapper/SILS/CCSDS_SILS_SCI_IF.h
@@ -1,23 +1,16 @@
 #ifndef CCSDS_SILS_SCI_IF_H_
 #define CCSDS_SILS_SCI_IF_H_
 
-#include <Windows.h>
+#include <stddef.h>
+#include "sils_sci_if.h"
 
-// ZEUS SILSからのコピー
-class SCIComPort
-{
-public:
-  SCIComPort(void);
-  ~SCIComPort(void);
 
-  int Send(unsigned char* buffer, size_t length, size_t offset);
-  int Receive(unsigned char* buffer, size_t length, size_t offset);
-
-private:
-  HANDLE myHComPort_;
-  DCB config_;
-  bool init_success;
-};
+// 最初だけ初期化して、プログラム終了時にポートを閉じるようにしたい
+#ifdef WIN32
+static SCIComPort SILS_SCI_IF_sci_com_(11);
+#else
+static SCIComPort SILS_SCI_IF_sci_com_(1);
+#endif
 
 int SILS_SIC_IF_init();
 int SILS_SIC_IF_TX(unsigned char* data_v, int count);
